requirements/ci.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    ./bin/compile_dependencies.sh
#
beautifulsoup4==4.9.0
    # via webtest
bleach==3.3.0
    # via -r requirements/base.txt
cairocffi==1.2.0
    # via
    #   -r requirements/base.txt
    #   cairosvg
    #   weasyprint
cairosvg==2.5.2
    # via
    #   -r requirements/base.txt
    #   weasyprint
certifi==2020.12.5
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.14.4
    # via
    #   -r requirements/base.txt
    #   cairocffi
    #   weasyprint
chardet==4.0.0
    # via
    #   -r requirements/base.txt
    #   requests
coverage==4.5.4
    # via -r requirements/test-tools.in
cssselect2==0.3.0
    # via
    #   -r requirements/base.txt
    #   cairosvg
    #   weasyprint
defusedxml==0.6.0
    # via
    #   -r requirements/base.txt
    #   cairosvg
    #   odfpy
diff-match-patch==20200713
    # via
    #   -r requirements/base.txt
    #   django-import-export
django-admin-tools==0.9.0
    # via -r requirements/base.txt
django-appconf==1.0.4
    # via
    #   -r requirements/base.txt
    #   django-compressor
django-autoslug==1.9.8
    # via -r requirements/base.txt
django-choices==1.7.1
    # via -r requirements/base.txt
django-ckeditor==6.0.0
    # via -r requirements/base.txt
django-compressor==2.1
    # via -r requirements/base.txt
django-extra-views==0.13.0
    # via -r requirements/base.txt
django-filter==2.4.0
    # via
    #   -r requirements/base.txt
    #   djangorestframework-filters
django-import-export==2.5.0
    # via -r requirements/base.txt
django-js-asset==1.2.2
    # via
    #   -r requirements/base.txt
    #   django-ckeditor
django-loginas==0.3.9
    # via -r requirements/base.txt
django-modeltranslation==0.16.1
    # via -r requirements/base.txt
django-precise-bbcode==1.2.13
    # via -r requirements/base.txt
django-rosetta==0.9.5
    # via -r requirements/base.txt
django-sendfile==0.3.11
    # via -r requirements/base.txt
django-sessionprofile==1.0
    # via -r requirements/base.txt
django-sniplates==0.7.0
    # via -r requirements/base.txt
django-solo==1.1.5
    # via -r requirements/base.txt
django-taggit==1.3.0
    # via -r requirements/base.txt
django-treebeard==4.3.1
    # via -r requirements/base.txt
django-webtest==1.9.7
    # via -r requirements/test-tools.in
django==2.2.20
    # via
    #   -r requirements/base.txt
    #   django-appconf
    #   django-choices
    #   django-extra-views
    #   django-filter
    #   django-import-export
    #   django-modeltranslation
    #   django-precise-bbcode
    #   django-rosetta
    #   django-sendfile
    #   django-sniplates
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
djangorestframework-filters==1.0.0.dev2
    # via -r requirements/base.txt
djangorestframework==3.11.2
    # via
    #   -r requirements/base.txt
    #   djangorestframework-filters
et-xmlfile==1.0.1
    # via
    #   -r requirements/base.txt
    #   openpyxl
factory-boy==2.12.0
    # via -r requirements/test-tools.in
faker==4.0.3
    # via factory-boy
freezegun==0.3.15
    # via -r requirements/test-tools.in
html2text==2020.1.16
    # via -r requirements/base.txt
html5lib==1.1
    # via
    #   -r requirements/base.txt
    #   weasyprint
idna==2.10
    # via
    #   -r requirements/base.txt
    #   requests
jdcal==1.4.1
    # via
    #   -r requirements/base.txt
    #   openpyxl
lxml==4.6.3
    # via -r requirements/base.txt
markuppy==1.14
    # via
    #   -r requirements/base.txt
    #   tablib
mysqlclient==2.0.3
    # via -r requirements/base.txt
odfpy==1.4.1
    # via
    #   -r requirements/base.txt
    #   tablib
openpyxl==3.0.5
    # via
    #   -r requirements/base.txt
    #   tablib
packaging==20.8
    # via
    #   -r requirements/base.txt
    #   bleach
pillow==8.2.0
    # via
    #   -r requirements/base.txt
    #   cairosvg
    #   django-precise-bbcode
    #   weasyprint
polib==1.1.0
    # via
    #   -r requirements/base.txt
    #   django-rosetta
psycopg2==2.8.6
    # via -r requirements/base.txt
pycparser==2.20
    # via
    #   -r requirements/base.txt
    #   cffi
pyparsing==2.4.7
    # via
    #   -r requirements/base.txt
    #   packaging
pyphen==0.10.0
    # via
    #   -r requirements/base.txt
    #   weasyprint
python-dateutil==2.8.1
    # via
    #   -r requirements/base.txt
    #   faker
    #   freezegun
python-decouple==3.3
    # via -r requirements/base.txt
python-dotenv==0.15.0
    # via -r requirements/base.txt
python-memcached==1.59
    # via -r requirements/base.txt
pytz==2020.5
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==5.4.1
    # via
    #   -r requirements/base.txt
    #   tablib
raven==6.10.0
    # via -r requirements/base.txt
rcssmin==1.0.6
    # via
    #   -r requirements/base.txt
    #   django-compressor
requests-mock==1.7.0
    # via -r requirements/test-tools.in
requests==2.25.1
    # via
    #   -r requirements/base.txt
    #   django-rosetta
    #   requests-mock
rjsmin==1.0.12
    # via
    #   -r requirements/base.txt
    #   django-compressor
six==1.15.0
    # via
    #   -r requirements/base.txt
    #   bleach
    #   django-choices
    #   django-extra-views
    #   django-modeltranslation
    #   django-rosetta
    #   freezegun
    #   html5lib
    #   python-dateutil
    #   python-memcached
    #   requests-mock
    #   webtest
sorl-thumbnail==12.7.0
    # via -r requirements/base.txt
soupsieve==2.0
    # via beautifulsoup4
sqlparse==0.4.1
    # via
    #   -r requirements/base.txt
    #   django
tablib[html,ods,xls,xlsx,yaml]==1.1.0
    # via
    #   -r requirements/base.txt
    #   django-import-export
tblib==1.6.0
    # via -r requirements/test-tools.in
text-unidecode==1.3
    # via faker
tinycss2==1.1.0
    # via
    #   -r requirements/base.txt
    #   cairosvg
    #   cssselect2
    #   weasyprint
urllib3==1.26.4
    # via
    #   -r requirements/base.txt
    #   requests
uwsgi==2.0.19.1
    # via -r requirements/base.txt
waitress==1.4.3
    # via webtest
weasyprint==52.2
    # via -r requirements/base.txt
webencodings==0.5.1
    # via
    #   -r requirements/base.txt
    #   bleach
    #   cssselect2
    #   html5lib
    #   tinycss2
webob==1.8.6
    # via webtest
webtest==2.0.34
    # via django-webtest
xlrd==1.2.0
    # via
    #   -r requirements/base.txt
    #   tablib
xlwt==1.3.0
    # via
    #   -r requirements/base.txt
    #   tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
